\name{ElpB1phenotype}
\alias{ElpB1phenotype}
\docType{data}
\title{
Ectopic wing vein formation phenotype caused by the EgfrElpB1.
}
\description{
Partial suppression of ectopic wing vein formation (in vivo fly phenotpye ) caused by the EgfrElpB1 in Cka heterozygous mutant backgrounds. The wing vein formation phenotype is classified as strong and notstrong. It is tested for three fly mutants.
}
\usage{data(ElpB1phenotype)}
\format{
  A data frame with 3 observations on the following 2 variables.
  \describe{
    \item{\code{strong}}{a numeric vector}
    \item{\code{notstrong}}{a numeric vector}
  }
}
\source{
Horn, Sandmann, Fischer, Huber, Boutros, Mapping of Signalling Networks through Synthetic Genetic Interaction Analysis by RNAi, Nature Methods, 2011, Figure 5f.
}
\examples{
data(ElpB1phenotype)
ElpB1phenotype
}
\keyword{datasets}
