\name{pathwayMembership}
\alias{pathwayMembership}
\docType{data}
\title{
Pathway membership of genes.
}
\description{
The membership of the tested genes in the four pathways JAK/STAT, RasMAPK, JNK, and p38. 
}
\usage{data(pathwayMembership)}
\format{
  The format is:
 chr "pathwayMembership"
}
\source{
Horn, Sandmann, Fischer, Huber, Boutros, Mapping of Signalling Networks through Synthetic Genetic Interaction Analysis by RNAi, Nature Methods, 2011, Figure S13.
}
\examples{
data(pathwayMembership)
head(pathwayMembership)
}
\keyword{datasets}
